Weird A* results on fourroute
runnerPA_AStarCL takes as command line arguments the network file and the sumocfg file. (Ideally we'd read the network file from the sumocfg file, but I haven't figured that out.)

Made a new map fourroute to show off the speed advantage from A*. But A* seems to prefer the longer outside routes. SUMO default routing prefers the inside routes, so it's not a problem with the network.